refactor console (#127)
* make console impl more simple and clear
