+++ console.py
@@ -11,122 +11,62 @@
   rlp, eth-rlp , hexbytes ... and relative projects
   @author: kentzhang
   @date: 2019-06
+
 """
-import argparse
 import sys
-import glob
-from client.gm_account import GM_Account
-from client.stattool import StatTool
-from client_config import client_config
-from eth_account.account import Account
-from eth_utils.hexadecimal import encode_hex
-from client.contractnote import ContractNote
-from eth_utils.crypto import CRYPTO_TYPE_GM
-import json
-import os
-from client.datatype_parser import DatatypeParser
-from eth_utils import to_checksum_address
+from console_utils.cmd_account import CmdAccount
+from console_utils.cmd_encode import CmdEncode
+from console_utils.cmd_transaction import CmdTransaction
+from console_utils.console_common import console_run_byname, contracts_dir
 from console_utils.precompile import Precompile
 from console_utils.rpc_console import RPCConsole
-from client.common import transaction_common
 from client.common import common
-from eth_abi.exceptions import InsufficientDataBytes
-
-from client.bcoserror import (
-    BcosError,
-    CompileError,
-    PrecompileError,
-    ArgumentsError,
-    BcosException,
-)
-from client.common.transaction_exception import TransactionException
-import argcomplete
-
-
-# --------------------------------------------------------------------------------------------
-# useful functions
-# --------------------------------------------------------------------------------------------
-
-
-def default_abi_file(contractname):
-    abi_file = contractname
-    if not abi_file.endswith(
-        ".abi"
-    ):  # default from contracts/xxxx.abi,if only input a name
-        abi_file = contracts_dir + "/" + contractname + ".abi"
-    return abi_file
+"""
+2020.10 重构，原来的实现很多塞在这文件里，有点乱，
+把命令分组为“账户account，交易transaction，内置合约precompile，读接口rpc，转码encode,
+每一组命令一个class，放到console_utils目录里
+把一些公共工具函数抽到console_common.py里
+在此文件主要是是解析输入，试图按配置和自动加载类结合的方式调取指定的命令实现.
+优化了下usage，可以指定分组的usage打印，如usage account,usage rpc,输出的信息比较少了，比较简洁易读
+"""
+cmd_mapping = dict()
+cmd_mapping["showaccount"] = ["cmd_account", "CmdAccount"]
+cmd_mapping["listaccount"] = ["cmd_account", "CmdAccount"]
+cmd_mapping["newaccount"] = ["cmd_account", "CmdAccount"]
+cmd_mapping["hex"] = ["cmd_encode", "CmdEncode"]
+cmd_mapping["decodehex"] = ["cmd_encode", "CmdEncode"]
+cmd_mapping["checkaddr"] = ["cmd_encode", "CmdEncode"]
+cmd_mapping["deploy"] = ["cmd_transaction", "CmdTransaction"]
+cmd_mapping["call"] = ["cmd_transaction", "CmdTransaction"]
+cmd_mapping["sendtx"] = ["cmd_transaction", "CmdTransaction"]
+cmd_mapping["deploylast"] = ["cmd_transaction", "CmdTransaction"]
+cmd_mapping["deploylog"] = ["cmd_transaction", "CmdTransaction"]
 
 
-def fill_params(params, paramsname):
-    index = 0
-    result = dict()
-    for name in paramsname:
-        result[name] = params[index]
-        index += 1
-    return result
-
-
-def print_receipt_logs_and_txoutput(client, receipt, contractname, parser=None):
-    print("INFO >>  receipt logs : ")
-    # 解析receipt里的log
-    if parser is None and len(contractname) > 0:
-        parser = DatatypeParser(default_abi_file(contractname))
-    logresult = parser.parse_event_logs(receipt["logs"])
-    i = 0
-    # print(json.dumps(logresult,indent=4))
-    for log in logresult:
-        if "eventname" in log:
-            i = i + 1
-            print(
-                "{}): log name: {} , data: {} , topic: {}".format(
-                    i, log["eventname"], log["eventdata"], log["topic"]
-                )
-            )
-    inputdetail = print_parse_transaction(receipt, "", parser)
-    # 解析该交易在receipt里输出的output,即交易调用的方法的return值
-    outputresults = parser.parse_receipt_output(inputdetail["name"], receipt["output"])
-    common.print_tx_result(outputresults)
-
-
-def print_parse_transaction(txReceipt, contractname, parser=None):
-    if "input" not in txReceipt:
-        return
-    if parser is None:
-        parser = DatatypeParser(default_abi_file(contractname))
-    inputdata = txReceipt["input"]
-    inputdetail = parser.parse_transaction_input(inputdata)
-    print("INFO >> transaction hash : ", txReceipt["transactionHash"])
-    print("tx input data detail:\n {}".format(inputdetail))
-    return inputdetail
-
-
-def check_result(result):
+def usage(inputparams=[]):
     """
-    check result
+    print usage
     """
-    if isinstance(result, dict) and "error" in result.keys():
-        return True
-    return False
+    print("FISCO BCOS 2.0 @python-SDK Usage:")
 
+    if len(inputparams) == 0:
+        module = "all"
+    else:
+        module = inputparams[0]
+    if module == 'all' or module == 'account':
+        CmdAccount.usage()
+    if module == 'all' or module == 'transaction':
+        CmdTransaction.usage()
+    if module == 'all' or module == 'encode':
+        CmdEncode.usage()
+    if module == 'all' or module == 'precompile':
+        Precompile.usage()
+    if module == 'all' or module == 'rpc':
+        RPCConsole.usage()
 
-def get_validcmds():
-    """
-    get valid cmds
-    """
-    validcmds = [
-        "showaccount",
-        "newaccount",
-        "deploylast",
-        "deploylog",
-        "deploy",
-        "call",
-        "sendtx",
-        "list",
-        "txinput",
-        "checkaddr",
-        "usage",
-    ]
-    return validcmds
+    print('''------------------------
+输入 : 'usage [module]' 查看指定控制台模块的指令(更加简洁):
+[ account , rpc , transaction ,  precompile , encode ] ''')
 
 
 def check_cmd(cmd, validcmds):
@@ -138,530 +78,60 @@ def check_cmd(cmd, validcmds):
     return True
 
 
-def printusage(usagemsg, precompile=None):
+def get_validcmds():
     """
-    print usage
+    get valid cmds
     """
-    print("FISCO BCOS 2.0 @python-SDK Usage:")
-    index = 0
-    for msg in usagemsg:
-        index += 1
-        print("{}\n".format(msg))
-    if precompile is None:
-        return
-    precompile.print_cns_usage(True)
-    precompile.print_consensus_usage(True)
-    precompile.print_sysconfig_usage(True)
-    precompile.print_all_permission_usage()
-    RPCConsole.print_rpc_usage()
-
-
-def usage(client_config):
-    """
-    generate usage list
-    """
-    usagemsg = []
-    usagemsg.append(
-        (
-            """newaccount [name] [password] [save]
-        创建一个新帐户，参数为帐户名(如alice,bob)和密码
-        结果加密保存在配置文件指定的帐户目录 *如同目录下已经有同名帐户文件，旧文件会复制一个备份
-        如输入了"save"参数在最后，则不做询问直接备份和写入
-        create a new account ,save to :[{}] (default) ,
-        the path in client_config.py:[account_keyfile_path]
-        if account file has exist ,then old file will save to a backup
-        if "save" arg follows,then backup file and write new without ask
-        the account len should be limitted to 240"""
-        ).format(client_config.account_keyfile_path)
-    )
-    usagemsg.append(
-        """showaccount [name] [password]
-        指定帐户名字(不带后缀)和密码，打开配置文件里默认账户文件路径下的[name].keystore文件，打印公私钥和地址
-        """
-    )
-    usagemsg.append(
-        """deploy [contract_name] [save]:
-        部署合约, 如给出'save'参数，新地址会写入本地记录文件,
-        eg: deploy SimpleInfo"""
-    )
-
-    usagemsg.append(
-        """call [contractname] [address] [func]  [args...]
-        call合约的一个只读接口,解析返回值
-        call a constant function of contract and get the returns
-        eg: call SimpleInfo 0xF2c07c98a6829aE61F3cB40c69f6b2f035dD63FC getbalance1 11
-        if address is "last" ,then load last address from :{}
-        eg: call SimpleInfo last getall
-        """.format(
-            client_config.contract_info_file
-        )
-    )
-
-    usagemsg.append(
-        """sendtx [contractname]  [address] [func] [args...]
-        发送交易调用指定合约的接口，交易如成功，结果会写入区块和状态
-        send transaction,will commit to blockchain if success
-        eg: sendtx SimpleInfo 0xF2c07c98a6829aE61F3cB40c69f6b2f035dD63FC
-        set alice 100 0xF2c07c98a6829aE61F3cB40c69f6b2f035dD63FC
-        if address is "last" ,then load last address from :{}
-        eg: sendtx SimpleInfo last set 'test' 100 '0xF2c07c98a6829aE61F3cB40c69f6b2f035dD63FC'
-        """.format(
-            client_config.contract_info_file
-        )
-    )
-
-    usagemsg.append(
-        """list
-        列出所有支持的get接口名和参数
-        list: list all  getcmds  has implemented
-        (getBlock...getTransaction...getReceipt..getOthers)"""
-    )
-
-    usagemsg.append(
-        """txinput [contractname] [inputdata(in hex string)]
-        复制一段来自transaction的inputdata(十六进制字符串)，指定合约名，则可以自动解析（合约的abi文件应存在指定目录下）
-        parse the transaction input data by  contractname，eg: txinput SimpleInfo [txinputdata]"""
-    )
-
-    usagemsg.append(
-        """checkaddr [address]
-        将普通地址转为自校验地址,自校验地址使用时不容易出错
-        change address to checksum address according EIP55:
-        to_checksum_address: 0xf2c07c98a6829ae61f3cb40c69f6b2f035dd63fc
-        -> 0xF2c07c98a6829aE61F3cB40c69f6b2f035dD63FC
-        """
-    )
-    return usagemsg
-
-
-def get_functions_by_contract(contract_name):
-    """
-    get functions according to contract_name
-    """
-    data_parser = DatatypeParser(default_abi_file(contract_name))
-    return [*data_parser.func_abi_map_by_name.keys()]
-
-
-def list_address(contract_name):
-    """
-    get address according to contract_name
-    """
-    return ContractNote.get_contract_addresses(contract_name)
-
-
-def list_api(file_pattern):
-    """
-    return list according to file_pattern
-    """
-    file_list = [f for f in glob.glob(file_pattern)]
-    targets = []
-    for file in file_list:
-        targets.append(os.path.basename(file).split(".")[0])
-    return targets
-
-
-def list_contracts():
-    """
-    list all contractname for call
-    """
-    return list_api(contracts_dir + "/*.sol")
-
-
-def list_accounts():
-    """
-    list all accounts
-    """
-    return list_api("bin/accounts/*.keystore")
-
-
-# get supported command
-Precompile.define_functions()
-RPCConsole.define_commands()
-validcmds = get_validcmds() + RPCConsole.get_all_cmd() + Precompile.get_all_cmd()
-contracts_dir = "contracts"
-
-
-def completion(prefix, parsed_args, **kwargs):
-    """
-    complete the shell
-    """
-    if parsed_args.cmd is None:
-        return validcmds
-    # deploy contract
-    if parsed_args.cmd[0] == "deploy":
-        return list_contracts()
-
-    # call and sendtx
-    # warn(parsed_args)
-    if parsed_args.cmd[0] == "call" or parsed_args.cmd[0] == "sendtx":
-        # only list the contract name
-        if len(parsed_args.cmd) == 1:
-            return list_contracts()
-        # list the contract address
-        if len(parsed_args.cmd) == 2:
-            return list_address(parsed_args.cmd[1])
-        # list functions
-        if len(parsed_args.cmd) == 3:
-            return get_functions_by_contract(parsed_args.cmd[1])
-
-    # call showaccount
-    if parsed_args.cmd[0] == "showaccount":
-        return list_accounts()
-
-    # registerCNS [contract_name] [contract_address] [contract_version]
-    if parsed_args.cmd[0] == "registerCNS":
-        # list contract name
-        if len(parsed_args.cmd) == 1:
-            return list_contracts()
-        # list contract address
-        if len(parsed_args.cmd) == 2:
-            return list_address(parsed_args.cmd[1])
-    # queryCNSByName [contract_name]
-    if parsed_args.cmd[0] == "queryCNSByName":
-        # list contract name
-        if len(parsed_args.cmd) == 1:
-            return list_contracts()
-    # queryCNSByNameAndVersion [contract_name] [contract_version]
-    if parsed_args.cmd[0] == "queryCNSByNameAndVersion":
-        if len(parsed_args.cmd) == 1:
-            return list_contracts()
-    # sysconfig
-    if (
-        parsed_args.cmd[0] == "setSystemConfigByKey"
-        or parsed_args.cmd[0] == "getSystemConfigByKey"
-    ):
-        return ["tx_count_limit", "tx_gas_limit"]
-    return []
+    cmds = []
+    for key in cmd_mapping:
+        cmds.append(key)
+    return cmds
 
 
 def parse_commands(argv):
-    """
-    parse the input command
-    """
-    # 首先创建一个ArgumentParser对象
-    parser = argparse.ArgumentParser(description="FISCO BCOS 2.0 lite client @python")
-    parsed_args = argparse.Namespace()
-    cmd = parser.add_argument("cmd", nargs="+", help="the command for console")  # 添加参数
-    cmd.completer = completion
-
-    argcomplete.autocomplete(parser)
-    args = parser.parse_args()
-
-    print("\nINFO >> user input : {}\n".format(args.cmd))
-    cmd = args.cmd[0]
-    inputparams = args.cmd[1:]
+    if len(argv) == 0:
+        cmd = "usage"
+        inputparams = []
+    else:
+        cmd = argv[0]
+        inputparams = argv[1:]
     return cmd, inputparams
 
 
+Precompile.define_functions()
+RPCConsole.define_commands()
+validcmds = get_validcmds() + RPCConsole.get_all_cmd() + Precompile.get_all_cmd() + ["usage"]
+
+
 def main(argv):
-    try:
-        usagemsg = usage(client_config)
-        cmd, inputparams = parse_commands(argv)
-        precompile = Precompile(cmd, inputparams, contracts_dir + "/precompile")
-        # check cmd
-        valid = check_cmd(cmd, validcmds)
-        if valid is False:
-            printusage(usagemsg, precompile)
-            return
-        # try to callback cns precompile
-        precompile.call_cns()
-        # try to callback consensus precompile
-        precompile.call_consensus()
-        # try to callback config precompile
-        precompile.call_sysconfig_precompile()
-        # try to callback permission precompile
-        precompile.call_permission_precompile()
-        # try to callback crud precompile
-        precompile.call_crud_precompile()
-        # try to callback rpc functions
-        rpcConsole = RPCConsole(cmd, inputparams, contracts_dir)
-        rpcConsole.executeRpcCommand()
-
-        def show_gm_account(name, password):
-            account = GM_Account()
-            keyfile = "{}/{}.json".format(client_config.account_keyfile_path, name)
-            account.load_from_file(keyfile, password)
-            print("load account from file: ", keyfile)
-            print(account.getdetail())
-
-        def show_ecdsa_account(name, password):
-
-            keyfile = "{}/{}.keystore".format(client_config.account_keyfile_path, name)
-            # the account doesn't exists
-            if os.path.exists(keyfile) is False:
-                raise BcosException("account {} doesn't exists".format(name))
-            print(
-                "show account : {}, keyfile:{} ,password {}  ".format(
-                    name, keyfile, password
-                )
-            )
-            try:
-                with open(keyfile, "r") as dump_f:
-                    keytext = json.load(dump_f)
-                    stat = StatTool.begin()
-                    privkey = Account.decrypt(keytext, password)
-                    stat.done()
-                    print("decrypt use time : %.3f s" % (stat.time_used))
-                    ac2 = Account.from_key(privkey)
-                    print("address:\t", ac2.address)
-                    print("privkey:\t", encode_hex(ac2.key))
-                    print("pubkey :\t", ac2.publickey)
-                    print("\naccount store in file: [{}]".format(keyfile))
-                    print("\n**** please remember your password !!! *****")
-            except Exception as e:
-                raise BcosException(
-                    ("load account info for [{}] failed," " error info: {}!").format(
-                        name, e
-                    )
-                )
-
-        if cmd == "showaccount":
-            # must be 2 params
-            common.check_param_num(inputparams, 2, True)
-            name = inputparams[0]
-            password = inputparams[1]
-            if client_config.crypto_type == CRYPTO_TYPE_GM:
-                show_gm_account(name, password)
-            else:
-                show_ecdsa_account(name, password)
-
-        def create_gm_account(name, password, forcewrite):
-            keyfile = "{}/{}.json".format(client_config.account_keyfile_path, name)
-            if not os.path.exists(keyfile):  # 如果默认文件不存在，直接写
-                forcewrite = True
-            else:
-                forcewrite = common.backup_file(keyfile)  # 如果备份失败，不要覆盖写
-
-            account = GM_Account()
-            account.create()
-            if forcewrite:
-                account.save_to_file(keyfile, password)
-            print("account created:")
-            print(account.getdetail())
-            if forcewrite:
-                print("account save to :", keyfile)
-
-        def create_ecdsa_account(name, password, forcewrite):
-            ac = Account.create(password)
-            print("new address :\t", ac.address)
-            print("new privkey :\t", encode_hex(ac.key))
-            print("new pubkey :\t", ac.publickey)
-
-            stat = StatTool.begin()
-            kf = Account.encrypt(ac.privateKey, password)
-            stat.done()
-            print("encrypt use time : %.3f s" % (stat.time_used))
-            keyfile = "{}/{}.keystore".format(client_config.account_keyfile_path, name)
-            print("save to file : [{}]".format(keyfile))
-            forcewrite = False
-            if not os.access(keyfile, os.F_OK):  # 默认的账号文件不存在，就强行存一个
-                forcewrite = True
-            else:
-                # old file exist,move to backup file first
-                if len(inputparams) == 3 and inputparams[2] == "save":
-                    forcewrite = True
-                else:
-                    forcewrite = common.backup_file(keyfile)  # 如果备份文件不成功，就不要覆盖写了
-            if forcewrite:
-                with open(keyfile, "w") as dump_f:
-                    json.dump(kf, dump_f)
-                    dump_f.close()
-            print(">>-------------------------------------------------------")
-            print(
-                "INFO >> read [{}] again after new account,address & keys in file:".format(
-                    keyfile
-                )
-            )
-            with open(keyfile, "r") as dump_f:
-                keytext = json.load(dump_f)
-                stat = StatTool.begin()
-                privkey = Account.decrypt(keytext, password)
-                stat.done()
-                print("decrypt use time : %.3f s" % (stat.time_used))
-                ac2 = Account.from_key(privkey)
-                print("address:\t", ac2.address)
-                print("privkey:\t", encode_hex(ac2.key))
-                print("pubkey :\t", ac2.publickey)
-                print("\naccount store in file: [{}]".format(keyfile))
-                print("\n**** please remember your password !!! *****")
-                dump_f.close()
-
-        if cmd == "newaccount":
-            common.check_param_num(inputparams, 2, False)
-            name = inputparams[0]
-            max_account_len = 240
-            if len(name) > max_account_len:
-                common.print_info(
-                    "WARNING",
-                    "account name should no more than {}".format(max_account_len),
-                )
-                sys.exit(1)
-            password = inputparams[1]
-            forcewrite = False
-            if len(inputparams) == 3 and inputparams[2] == "save":
-                forcewrite = True
-            print("starting : {} {}  , if save:{}".format(name, password, forcewrite))
-            if client_config.crypto_type == CRYPTO_TYPE_GM:
-                create_gm_account(name, password, forcewrite)
-            else:
-                create_ecdsa_account(name, password, forcewrite)
-
-        if cmd == "deploylast":
-            contracts = ContractNote.get_last_contracts()
-            for name in contracts:
-                print("{} -> {}".format(name, contracts[name]))
-
-        if cmd == "deploylog":
-            historys = ContractNote.get_history_list()
-            for address in historys:
-                print("{} -> {} ".format(address, historys[address]))
-
-        # --------------------------------------------------------------------------------------------
-        # console cmd entity
-        # --------------------------------------------------------------------------------------------
-        if cmd == "deploy":
-            """deploy abi bin file"""
-            # must be at least 2 params
-            common.check_param_num(inputparams, 1)
-            contractname = inputparams[0].strip()
-            gasPrice = 30000000
-            # need save address whether or not
-            needSaveAddress = False
-            args_len = len(inputparams)
-            if inputparams[-1] == "save":
-                needSaveAddress = True
-                args_len = len(inputparams) - 1
-            # get the args
-            fn_args = inputparams[1:args_len]
-            tx_client = transaction_common.TransactionCommon(
-                "", contracts_dir, contractname
-            )
-            result = tx_client.send_transaction_getReceipt(
-                None, fn_args, gasPrice, True
-            )[0]
-            print("INFO >> client info: {}".format(tx_client.getinfo()))
-            print(
-                "deploy result  for [{}] is:\n {}".format(
-                    contractname, json.dumps(result, indent=4)
-                )
-            )
-            name = contractname
-            address = result["contractAddress"]
-            blocknum = int(result["blockNumber"], 16)
-            txhash = result["transactionHash"]
-            ContractNote.save_contract_address(name, address)
-            print("on block : {},address: {} ".format(blocknum, address))
-            if needSaveAddress is True:
-                ContractNote.save_address_to_contract_note(name, address)
-                print("address save to file: ", client_config.contract_info_file)
-            else:
-                print(
-                    """\nNOTE : if want to save new address as last
-                    address for (call/sendtx)\nadd 'save' to cmdline and run again"""
-                )
-            ContractNote.save_history(name, address, blocknum, txhash)
-
-        # --------------------------------------------------------------------------------------------
-        # console cmd entity
-        # --------------------------------------------------------------------------------------------
-        if cmd == "call" or cmd == "sendtx":
-            common.check_param_num(inputparams, 3)
-            paramsname = ["contractname", "address", "func"]
-            params = fill_params(inputparams, paramsname)
-            contractname = params["contractname"]
-            address = params["address"]
-            if address == "last":
-                address = ContractNote.get_last(contractname)
-                if address is None:
-                    sys.exit(
-                        "can not get last address for [{}],break;".format(contractname)
-                    )
-
-            tx_client = transaction_common.TransactionCommon(
-                address, contracts_dir, contractname
-            )
-            fn_name = params["func"]
-            fn_args = inputparams[3:]
-            print("INFO>> client info: {}".format(tx_client.getinfo()))
-            print(
-                "INFO >> {} {} , address: {}, func: {}, args:{}".format(
-                    cmd, contractname, address, fn_name, fn_args
-                )
-            )
-            if cmd == "call":
-                result = tx_client.call_and_decode(fn_name, fn_args)
-                print("INFO >> {} result:".format(cmd))
-                common.print_tx_result(result)
-            if cmd == "sendtx":
-                receipt = tx_client.send_transaction_getReceipt(fn_name, fn_args)[0]
-                data_parser = DatatypeParser(default_abi_file(contractname))
-                print("\n\nINFO >> from address:  {} ".format(tx_client.keypair.address))
-                # 解析receipt里的log 和 相关的tx ,output
-                print_receipt_logs_and_txoutput(tx_client, receipt, "", data_parser)
-        # --------------------------------------------------------------------------------------------
-        # console cmd entity
-        # --------------------------------------------------------------------------------------------
-        if cmd == "list":
-            RPCConsole.print_rpc_usage()
-            print(
-                "--------------------------------------------------------------------"
-            )
-
-        # --------------------------------------------------------------------------------------------
-        # console cmd entity
-        # --------------------------------------------------------------------------------------------
-        if cmd == "txinput":
-            contractname = inputparams[0]
-            inputdata = inputparams[1]
-            abi_path = default_abi_file(contractname)
-            if os.path.isfile(abi_path) is False:
-                raise BcosException(
-                    "execute {} failed for {} doesn't exist".format(cmd, abi_path)
-                )
-            try:
-                dataParser = DatatypeParser(abi_path)
-                # print(dataParser.func_abi_map_by_selector)
-                result = dataParser.parse_transaction_input(inputdata)
-                print("\nabifile : ", default_abi_file(contractname))
-                print("parse result: {}".format(result))
-            except Exception as e:
-                raise BcosException("execute {} failed for reason: {}".format(cmd, e))
-
-        # --------------------------------------------------------------------------------------------
-        # console cmd entity
-        # --------------------------------------------------------------------------------------------
-        if cmd == "checkaddr":
-            address = inputparams[0]
-            result = to_checksum_address(address)
-            print("{} -->\n{}".format(address, result))
-
-        # --------------------------------------------------------------------------------------------
-        # console cmd entity
-        # --------------------------------------------------------------------------------------------
-        if cmd == "usage":
-            printusage(usagemsg, precompile)
-    except TransactionException as e:
-        common.print_error_msg(cmd, e)
-    except PrecompileError as e:
-        common.print_error_msg(cmd, e)
-    except BcosError as e:
-        common.print_error_msg(cmd, e)
-    except CompileError as e:
-        common.print_error_msg(cmd, e)
-    except ArgumentsError as e:
-        common.print_error_msg(cmd, e)
-    except BcosException as e:
-        common.print_error_msg(cmd, e)
-    except ValueError as e:
-        common.print_error_msg(cmd, e)
-    except InsufficientDataBytes as e:
-        common.print_error_msg(cmd, e)
-    except Exception as e:
-        print("exception happened!")
-        import traceback
-
-        print(traceback.format_exc())
-        exit(-1)
+    cmd, inputparams = parse_commands(argv)
+    # check cmd
+    valid = check_cmd(cmd, validcmds)
+    if valid is False:
+        usage()
+        return
+    if cmd == "usage":
+        usage(inputparams)
+        return
+    if cmd in cmd_mapping:
+        (modulename, classname) = cmd_mapping[cmd]
+        console_run_byname(modulename, classname, cmd, inputparams)
+        return
+    print("cmd=", cmd)
+    precompile = Precompile(cmd, inputparams, contracts_dir + "/precompile")
+    # try to callback cns precompile
+    precompile.call_cns()
+    # try to callback consensus precompile
+    precompile.call_consensus()
+    # try to callback config precompile
+    precompile.call_sysconfig_precompile()
+    # try to callback permission precompile
+    precompile.call_permission_precompile()
+    # try to callback crud precompile
+    precompile.call_crud_precompile()
+    # try to callback rpc functions
+    rpcconsole = RPCConsole(cmd, inputparams, contracts_dir)
+    rpcconsole.executeRpcCommand()
 
 
 if __name__ == "__main__":
